feat: Update server #2
