+++ server/middlewares/header-middleware.js
@@ -1,0 +1,12 @@
+const setHeadersMiddleware = (req, res, next) => {
+  res.removeHeader('x-powered-by')
+  res.header('no-cache', 'Set-Cookie')
+  res.header('x-xss-protection', '1; mode=block')
+  res.header('x-frame-options', 'DENY')
+  res.header('x-content-type-options', 'nosniff')
+  next()
+}
+
+module.exports = {
+  setHeadersMiddleware
+}
